+++ Changelog.md
@@ -20,6 +20,7 @@ All notable changes to this project will be documented in this file!
 - Changed: The most recent Task action added is highlighted to distinquish it from all of the other actions.
 - Changed: The current Task action being added now prompts for the arguments immeidately.
 - Changed: 'Save Single Task' button renamed to 'Export Task', and 'Save Single Profile' to 'Export Profile'.
+- Changed: Project and Profile 'Save to Android' now saves the Project/Profile only to /Tasker/projects and /Tasker/projects, respectfully, on the Android device, since these can't be directly inserted into Tasker.
 
 ### Fixed